stop broadcast video before starting playback
The HbbTV specification requires that any broadcast video playback is
stopped in order to release the decoders and allow them to be used for a
HTML video element (HbbTV § A.2.1).

Change d7dff31a29bd made the code which did this conditional so that it
only happened some of the time.  It is not clear what bug this change
was intended to fix.  The original code called bindToCurrentChannel each
time through, restarting broadcast playback if it had been stopped
already.  My best guess is that this is what the original change was
intended to fix.